+++ src/options.rs
@@ -4,25 +4,25 @@ use structopt::StructOpt;
 #[derive(Debug, Clone, StructOpt)]
 pub struct Options {
     /// Shows only this number of top types.
-    #[structopt(short, long, default_value = "100")]
-    pub limit: usize,
+    #[structopt(short = "l", long)]
+    pub limit: Option<usize>,
     /// Prints top `limit` types in ascending order.
-    #[structopt(short, long)]
+    #[structopt(short = "r", long)]
     pub reverse: bool,
     /// Removes wrappers like `MaybeUninit`.
     #[structopt(short = "w", long)]
     pub remove_wrappers: bool,
     /// Hides types and fields with size less than this value.
-    #[structopt(short, long, default_value = "0")]
-    pub hide_less: usize,
+    #[structopt(short = "h", long)]
+    pub hide_less: Option<usize>,
     /// Sorts fields by size and removes paddings.
-    #[structopt(short, long)]
+    #[structopt(short = "s", long)]
     pub sort_fields: bool,
     /// Shows only types that match these patterns.
-    #[structopt(short, long)]
+    #[structopt(short = "f", long)]
     pub filter: Vec<Regex>,
     /// Excludes types that match these patterns.
-    #[structopt(short, long)]
+    #[structopt(short = "e", long)]
     pub exclude: Vec<Regex>,
     /// Shows only types that match these patterns and their children,
     /// heuristically.