+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - The `-h` parameter now filters also types, not only fields.
+- Remove a default value for `-l`, now output is unlimited by default.
 
 ### Fixed
 - Support nightly after 2024-03-22 ([#4]).